Implemented subvoxel pore desaturation
Compiles but not yet tested

This tracks and modifies only global subvoxel pore saturation; phases
are allowed to nucleate in void space that neighbors porous solids, but
this version does not keep track of localized subvoxel porosity on a
voxel-by-voxel basis. Instead, it is assumed that subvoxel pore water
redistributes among all porous solids in the system.